requirements.txt
fastapi
uvicorn
pydantic
numpy
openai
python-dotenv
google-generativeai
